修改Point中x,y的属性

+++ src/main/java/com/parkinglot_backend/dto/NavigationPoint.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class NavigationPoint {
-    private int startX;
-    private int startY;
-    private int endX;
-    private int endY;
+    private double startX;
+    private double startY;
+    private double endX;
+    private double endY;
 }
